Refactor syncUp&buttonClick to improve testing&readablity
We want to pass the target DOM elements &request instance to callback functions such as syncUp and Button click as arguments to help with their readablity&testing.

But setInterval& eventListener callback does not allow us to pass the function arguments explicitly. We use this anonymous function wrappers to bypass that