Add Panel wrapper for status panel DOM elements
We only initialize panel at window.onload by passing the actual DOM element pointers because that is called after DOM elements are loaded.

The constructor with method functions is initailized on the main level, with everything else.

Another plus to this approach is that we can write unit tests for the Panel easily.