Make users be able to press thebutton only once per session
Wehn the user clicks the button, we store that info in the Flask's session object and not allow the user to click it again.

This is supposed to simulate the functionality of each user being able to click only once which makes the reddit's button so captivating.

Of course, here there is no registered users so we cannot repeat that functionality completely. However having to close the browser to redo it again simulates it nicely enough for our purposes.